Move footnote context to DocumentBody
